fix(zsh): append `_zoxide_precmd` as a single-element array to hook
We should append `_zoxide_precmd` to `precmd_functions` within an array literal, not by itself.

This is important because...well, nothing works unless you append an *array* to another array, because that's what's expected in `zsh` and `bash`! Is it weird? Yes. But that's the type system we live with!