Allow use of cd to positions in dirstack for zsh
man zshbuiltins excerpt:
The  third  form  of  cd extracts an entry from the directory stack, and
changes to that directory.  An argument of the form  `+n'  identifies  a
stack entry by counting from the left of the list shown by the dirs com‐
mand, starting with zero.  An argument of the form `-n' counts from  the
right.  If the PUSHD_MINUS option is set, the meanings of `+' and `-' in
this context are swapped.  If the POSIX_CD option is set, this  form  of
cd is not recognised and will be interpreted as the first form.