improve safety while using variables with utilities by passing `--`
This pattern *is* POSIX defined (see [here](https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html#tag_12_02)):
```
Guideline 10:
The first -- argument that is not an option-argument should be accepted as a delimiter indicating the end of options. Any following arguments should be treated as operands, even if they begin with the '-' character.
```

This ensures that odd values for _bin_name, _bin_dir, and others aren't treated as options to the command and cause a failure.