support autocd option (#695)
* support autocd option

In bash, when autocd option is set and user enters a directory name as a
command, it results in a very specific call to cd:

cd -- [directory name]

zoxide's directory changing function passes all arguments to __zoxide_z as is,
including the "--" first argument. By detecting this and skipping the first
argument changing directory works with autocd set.

* remove directory check

just skip the first argument and pass the rest as is. checking for directory
breaks cd'ing to symlinked directories

* undo some of the hackery

tests are failing