Care about terminals not responding to DSR in Bash integration
The current method of using the DSR request (\e[5n) and response
(\e[0n) for the operating status relies on the assumption that the
terminal supports the DSR request for the operating status and always
responds with the good status (\e[0n) instead of the bad one (\e[3n).
However, the terminal may not support \e[5n or may return another
status in principle.

Moreover, when a non-trivial shell framework is involved, the DSR
response might be processed by an unintended part.  For example, when
ble.sh (https://github.com/akinomyoga/ble.sh) is loaded in the Bash
session, since we use "\builtin bind" instead of "bind" replaced by
ble.sh, the response would be processed by Readline (which is
deactivated by ble.sh) instead of ble.sh (i.e., the line editor in
action).  As a result, the attempt to replace the command line will
fail.

As a fallback for such a situation, it is still worth inserting the
target directory through the command line.  This patch restores an
older logic to insert the completion.

This patch provides a fallback to the issue reported in the following
places:

[1] https://github.com/ajeetdsouza/zoxide/issues/1043#issuecomment-2845790391
[2] https://github.com/akinomyoga/ble.sh/issues/580#issuecomment-2845818748