Use "bind" instead of "\builtin bind"
The ble.sh framework would be broken if the keybinding is directly
changed using "bind" because the entire ble.sh line editor is built on
top of "builtin bind -x".

This patch replaces "\builtin bind" with "bind".  This patch also
explicitly redirects the output of "printf '\e[5n'" to /dev/tty
because ble.sh's Zoxide integration suppresses stdout in executing the
Zoxide completion.

Note: By default, ble.sh discards unrecognized CSI sequences, so one
needs to change the setting to receive the DSR response \e[0n:

    # In ~/.blerc or ~/.config/blesh/init.sh

    bleopt decode_error_cseq_discard=

In the latest version of ble.sh (>= adf53ed3), the above setting is
unnecessary because \e[0n was added as a distinct key "dsr0".